feat:Add Distributed Tracing support for Azure Functions HTTPTrigger.
